requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
dotenv==0.9.9
eventlet==0.40.4
Flask==3.1.2
Flask-Login==0.6.3
Flask-SocketIO==5.6.0
Flask-WTF==1.2.2
greenlet==3.3.1
gunicorn==25.0.3
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
Werkzeug==3.1.5
wsproto==1.3.2
WTForms==3.2.1
